+++ src/main/java/fr/oukilson/backend/entity/Game.java
@@ -1,19 +1,26 @@
-package fr.oukilson.backend.dtos;
+package fr.oukilson.backend.entity;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
+import javax.persistence.*;
 
-@Data
+@Entity
+@Table(name = "game")
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class GameDTO {
+public class Game {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
     private String uuid;
     private String name;
     private Integer minPlayer;
     private Integer maxPlayer;
+    // Duration in minutes
     private Integer minTime;
     private Integer maxTime;
     private Integer minAge;
     private String creatorName;
+
 }
