+++ src/main/java/fr/oukilson/backend/repository/GameRepository.java
@@ -1,6 +1,6 @@
 package fr.oukilson.backend.repository;
 
-import fr.oukilson.backend.entities.Game;
+import fr.oukilson.backend.entity.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
